Created regression test for morton builder clamp
