+++ tests/CMakeLists.txt
@@ -39,6 +39,8 @@ FOREACH(xml ${PRIMITIVE_TESTS})
   ENDFOREACH()
 ENDFOREACH()
 
+  ADD_SUBDIRECTORY(regression)
+
 
 IF (EMBREE_TESTING_INSTALL_TESTS)
   # test resources
